src/main/java/ru/netology/manager/TicketManager.java
package ru.netology.manager;

import ru.netology.repository.TicketRepository;
import ru.netology.ticket.Ticket;

import java.util.Arrays;

public class TicketManager {
    private TicketRepository repository;

    public TicketManager(TicketRepository repository) {
        this.repository = repository;
    }

    public void add(Ticket ticket) {
        repository.save(ticket);
    }

    public Ticket[] searchBy(String textFrom, String textTo) {
        Ticket[] result = new Ticket[0];
        for (Ticket ticket : repository.findAll()) {
            if (matchesFrom(ticket, textFrom) && matchesTo(ticket, textTo)) {
                Ticket[] tmp = new Ticket[result.length + 1];
                for (int i = 0; i < result.length; i++) {
                    tmp[i] = result[i];
                }
                tmp[tmp.length - 1] = ticket;
                result = tmp;
                Arrays.sort(result);
            }
        }
        return result;
    }

    public boolean matchesFrom(Ticket ticket, String searchFrom) {
        return ticket.getFrom().contains(searchFrom);
    }

    public boolean matchesTo(Ticket ticket, String searchTo) {
        return ticket.getTo().contains(searchTo);

    }
}
